refactor: 역할에 맞게 패키지 구조를 수정하라
- input 객체 및  input validator 객체는 모두 input 의 행위에 해당하므로 input package 에 위치한다
