+++ src/main/java/racingcar/input/validator/impl/CarNamesOverlapValidator.java
@@ -1,11 +1,11 @@
-package racingcar.domain.validator.impl;
+package racingcar.input.validator.impl;
 
 import java.util.Arrays;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 import racingcar.domain.ExceptionType;
-import racingcar.domain.validator.InputValidator;
+import racingcar.input.validator.InputValidator;
 
 public class CarNamesOverlapValidator implements InputValidator {
     @Override
