+++ src/test/java/racingcar/domain/CarMoveTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mockStatic;
@@ -11,8 +11,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import org.mockito.MockedStatic;
-import racingcar.domain.Car;
-import racingcar.domain.MoveBehavior;
 
 public class CarMoveTest {
     private MockedStatic<Randoms> randoms;
